Check data harmonization

Functions to check that the input data frame (trait table or allometry table) has the appropriate format and data for model parameter filling.

Usage

check_harmonized_trait(x, verbose = TRUE)

check_harmonized_trait_dir(harmonized_trait_path, verbose = TRUE)

check_harmonized_allometry(x, verbose = TRUE)

check_harmonized_allometry_dir(harmonized_allometry_path, verbose = TRUE)

Arguments

x

A data frame with harmonized trait data or harmonized allometry data

verbose

A logical flag indicating extra console output (information alerts and check success)

harmonized_trait_path

The path to harmonized trait data files (.rds or .csv format).

harmonized_allometry_path

The path to harmonized allometry data files (.rds or .csv format).

Value

An (invisible) logical indicating whether the data frame is acceptable or not.

Details

Function check_harmonized_trait() checks that the input data frame conforms to the following requirements:

Has columns called originalName, acceptedName, acceptedNameAuthorship, family, genus, specificEpithet, and taxonRank, as returned by function harmonize_taxonomy_WFO.

The names of the remaining columns are "Units", "Reference" (data source citation), "DOI" (data source DOI), "OriginalReference" (original data source citation),"OriginalDOI" (original data source DOI), "Priority" or a valid trait name according to the notation required in HarmonizedTraitDefinition.

For trait columns, their values conform to the required type in HarmonizedTraitDefinition.

Trait columns contain non-missing data.

Numeric trait columns do not contain values beyond expected ranges (when defined).

Columns "Reference" and "DOI" refer to the trait data source. In the case of a trait data base, the original sources can be specified in "OriginalReference" and "OriginalDOI". Function check_harmonized_allometry() checks that the input data frame conforms to the following requirements:

Has columns called originalName, acceptedName, acceptedNameAuthorship, family, genus, specificEpithet, and taxonRank, as returned by function harmonize_taxonomy_WFO.

Has columns called "Response" and "Equation". Columns "Priority", "Reference" and "DOI" are recommended.

Column "Response" identifies the response variable.

Column "ResponseDescription" contains a longer description of the response variable.

Columns "Predictor1", "Predictor2", ... identify predictor variables.

Columns "PredictorDescription1", "PredictorDescription2", ... contain longer descriptions of predictor variables.

Allometry coefficient columns should be named "a", "c", "c", ..., "f" and contain numeric values.

See also