Initializes species parameter table

Creates an empty plant parameter table for medfate, populating taxonomic information if desired.

init_spparams(
  sp_names,
  accepted_names = NULL,
  fill_taxonomy = TRUE,
  complete_rows = TRUE,
  sort = TRUE,
  verbose = FALSE
)

Arguments

sp_names

Vector of plant names, either taxon names or arbitrary species group names.

accepted_names

Vector of accepted taxon names of the same length of sp_names.

fill_taxonomy

Boolean flag to indicate that taxonomic information should be filled (retrieved from GBIF using package taxize).

complete_rows

Boolean flag to indicate that extra rows should be added for cited species/genera (if fill_taxonomy = TRUE).

sort

Boolean flag to force sorting in ascending order by Name.

verbose

A boolean flag to indicate extra console output.

Value

A data frame with empty species parameter values suitable for medfate simulations. The data frame will normally contain more rows than sp_names because of arguments fill_taxonomy and complete_rows are set to TRUE by default.

Details

Taxonomic information is retrieved using functions in package taxize and GBIF as data source.

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

# \donttest{
# Simple example with two species
sp_names <- c("Salvia rosmarinus", "Pinus contorta")
init_spparams(sp_names)
#>                Name SpIndex      AcceptedName           Species  Genus
#> 1             Pinus       0             Pinus              <NA>  Pinus
#> 2    Pinus contorta       1    Pinus contorta    Pinus contorta  Pinus
#> 3            Salvia       2            Salvia              <NA> Salvia
#> 4 Salvia rosmarinus       3 Salvia rosmarinus Salvia rosmarinus Salvia
#>      Family    Order      Group GrowthForm LifeForm LeafShape LeafSize
#> 1  Pinaceae  Pinales Gymnosperm         NA       NA        NA       NA
#> 2  Pinaceae  Pinales Gymnosperm         NA       NA        NA       NA
#> 3 Lamiaceae Lamiales Angiosperm         NA       NA        NA       NA
#> 4 Lamiaceae Lamiales Angiosperm         NA       NA        NA       NA
#>   PhenologyType DispersalType Hmed Hmax Z50 Z95 fHDmin fHDmax a_ash b_ash a_bsh
#> 1            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 2            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 3            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 4            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#>   b_bsh a_btsh b_btsh cr BTsh a_fbt b_fbt c_fbt a_cr b_1cr b_2cr b_3cr c_1cr
#> 1    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 2    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 3    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 4    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#>   c_2cr a_cw b_cw a_bt b_bt LeafDuration t0gdd Sgdd Tbgdd Ssen Phsen Tbsen xsen
#> 1    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 2    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 3    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 4    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#>   ysen SLA LeafDensity WoodDensity FineRootDensity conduit2sapwood r635 pDead
#> 1   NA  NA          NA          NA              NA              NA   NA    NA
#> 2   NA  NA          NA          NA              NA              NA   NA    NA
#> 3   NA  NA          NA          NA              NA              NA   NA    NA
#> 4   NA  NA          NA          NA              NA              NA   NA    NA
#>   Al2As Ar2Al LeafWidth SRL RLD maxFMC minFMC LeafPI0 LeafEPS LeafAF StemPI0
#> 1    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 2    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 3    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 4    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#>   StemEPS StemAF SAV HeatContent LigninPercent LeafAngle LeafAngleSD
#> 1      NA     NA  NA          NA            NA        NA          NA
#> 2      NA     NA  NA          NA            NA        NA          NA
#> 3      NA     NA  NA          NA            NA        NA          NA
#> 4      NA     NA  NA          NA            NA        NA          NA
#>   ClumpingIndex gammaSWR alphaSWR kPAR  g Tmax_LAI Tmax_LAIsq Psi_Extract
#> 1            NA       NA       NA   NA NA       NA         NA          NA
#> 2            NA       NA       NA   NA NA       NA         NA          NA
#> 3            NA       NA       NA   NA NA       NA         NA          NA
#> 4            NA       NA       NA   NA NA       NA         NA          NA
#>   Exp_Extract WUE WUE_par WUE_co2 WUE_vpd Gswmin Gswmax Gs_Toptim Gs_Tsens
#> 1          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 2          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 3          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 4          NA  NA      NA      NA      NA     NA     NA        NA       NA
#>   Gs_P50 Gs_slope VCleaf_kmax VCleaf_P12 VCleaf_P50 VCleaf_P88 VCleaf_slope
#> 1     NA       NA          NA         NA         NA         NA           NA
#> 2     NA       NA          NA         NA         NA         NA           NA
#> 3     NA       NA          NA         NA         NA         NA           NA
#> 4     NA       NA          NA         NA         NA         NA           NA
#>   Kmax_stemxylem VCstem_P12 VCstem_P50 VCstem_P88 VCstem_slope Kmax_rootxylem
#> 1             NA         NA         NA         NA           NA             NA
#> 2             NA         NA         NA         NA           NA             NA
#> 3             NA         NA         NA         NA           NA             NA
#> 4             NA         NA         NA         NA           NA             NA
#>   VCroot_P12 VCroot_P50 VCroot_P88 VCroot_slope Vmax298 Jmax298 Nleaf Nsapwood
#> 1         NA         NA         NA           NA      NA      NA    NA       NA
#> 2         NA         NA         NA           NA      NA      NA    NA       NA
#> 3         NA         NA         NA           NA      NA      NA    NA       NA
#> 4         NA         NA         NA           NA      NA      NA    NA       NA
#>   Nfineroot WoodC RERleaf RERsapwood RERfineroot CCleaf CCsapwood CCfineroot
#> 1        NA    NA      NA         NA          NA     NA        NA         NA
#> 2        NA    NA      NA         NA          NA     NA        NA         NA
#> 3        NA    NA      NA         NA          NA     NA        NA         NA
#> 4        NA    NA      NA         NA          NA     NA        NA         NA
#>   RGRleafmax RGRsapwoodmax RGRcambiummax RGRfinerootmax SRsapwood SRfineroot
#> 1         NA            NA            NA             NA        NA         NA
#> 2         NA            NA            NA             NA        NA         NA
#> 3         NA            NA            NA             NA        NA         NA
#> 4         NA            NA            NA             NA        NA         NA
#>   RSSG MortalityBaselineRate SurvivalModelStep SurvivalB0 SurvivalB1
#> 1   NA                    NA                NA         NA         NA
#> 2   NA                    NA                NA         NA         NA
#> 3   NA                    NA                NA         NA         NA
#> 4   NA                    NA                NA         NA         NA
#>   SeedProductionHeight SeedMass SeedLongevity DispersalDistance DispersalShape
#> 1                   NA       NA            NA                NA             NA
#> 2                   NA       NA            NA                NA             NA
#> 3                   NA       NA            NA                NA             NA
#> 4                   NA       NA            NA                NA             NA
#>   ProbRecr MinTempRecr MinMoistureRecr MinFPARRecr RecrTreeDBH RecrTreeHeight
#> 1       NA          NA              NA          NA          NA             NA
#> 2       NA          NA              NA          NA          NA             NA
#> 3       NA          NA              NA          NA          NA             NA
#> 4       NA          NA              NA          NA          NA             NA
#>   RecrShrubHeight RecrTreeDensity RecrShrubCover RecrZ50 RecrZ95 RespFire
#> 1              NA              NA             NA      NA      NA       NA
#> 2              NA              NA             NA      NA      NA       NA
#> 3              NA              NA             NA      NA      NA       NA
#> 4              NA              NA             NA      NA      NA       NA
#>   RespDist RespClip IngrowthTreeDensity IngrowthTreeDBH
#> 1       NA       NA                  NA              NA
#> 2       NA       NA                  NA              NA
#> 3       NA       NA                  NA              NA
#> 4       NA       NA                  NA              NA

# Simple example with three species using synonyms and subspecies
sp_names <- c("Rosmarinus officinalis", "Pinus contorta", "Quercus ilex subsp. ilex")
accepted_names <- c("Salvia rosmarinus", "Pinus contorta", "Quercus ilex subsp. ilex")
init_spparams(sp_names, accepted_names)
#>                       Name SpIndex             AcceptedName           Species
#> 1                    Pinus       0                    Pinus              <NA>
#> 2           Pinus contorta       1           Pinus contorta    Pinus contorta
#> 3                  Quercus       2                  Quercus              <NA>
#> 4 Quercus ilex subsp. ilex       3 Quercus ilex subsp. ilex      Quercus ilex
#> 5   Rosmarinus officinalis       4        Salvia rosmarinus Salvia rosmarinus
#> 6                   Salvia       5                   Salvia              <NA>
#>     Genus    Family    Order      Group GrowthForm LifeForm LeafShape LeafSize
#> 1   Pinus  Pinaceae  Pinales Gymnosperm         NA       NA        NA       NA
#> 2   Pinus  Pinaceae  Pinales Gymnosperm         NA       NA        NA       NA
#> 3 Quercus  Fagaceae  Fagales Angiosperm         NA       NA        NA       NA
#> 4 Quercus  Fagaceae  Fagales Angiosperm         NA       NA        NA       NA
#> 5  Salvia Lamiaceae Lamiales Angiosperm         NA       NA        NA       NA
#> 6  Salvia Lamiaceae Lamiales Angiosperm         NA       NA        NA       NA
#>   PhenologyType DispersalType Hmed Hmax Z50 Z95 fHDmin fHDmax a_ash b_ash a_bsh
#> 1            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 2            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 3            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 4            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 5            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#> 6            NA            NA   NA   NA  NA  NA     NA     NA    NA    NA    NA
#>   b_bsh a_btsh b_btsh cr BTsh a_fbt b_fbt c_fbt a_cr b_1cr b_2cr b_3cr c_1cr
#> 1    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 2    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 3    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 4    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 5    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#> 6    NA     NA     NA NA   NA    NA    NA    NA   NA    NA    NA    NA    NA
#>   c_2cr a_cw b_cw a_bt b_bt LeafDuration t0gdd Sgdd Tbgdd Ssen Phsen Tbsen xsen
#> 1    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 2    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 3    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 4    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 5    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#> 6    NA   NA   NA   NA   NA           NA    NA   NA    NA   NA    NA    NA   NA
#>   ysen SLA LeafDensity WoodDensity FineRootDensity conduit2sapwood r635 pDead
#> 1   NA  NA          NA          NA              NA              NA   NA    NA
#> 2   NA  NA          NA          NA              NA              NA   NA    NA
#> 3   NA  NA          NA          NA              NA              NA   NA    NA
#> 4   NA  NA          NA          NA              NA              NA   NA    NA
#> 5   NA  NA          NA          NA              NA              NA   NA    NA
#> 6   NA  NA          NA          NA              NA              NA   NA    NA
#>   Al2As Ar2Al LeafWidth SRL RLD maxFMC minFMC LeafPI0 LeafEPS LeafAF StemPI0
#> 1    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 2    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 3    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 4    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 5    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#> 6    NA    NA        NA  NA  NA     NA     NA      NA      NA     NA      NA
#>   StemEPS StemAF SAV HeatContent LigninPercent LeafAngle LeafAngleSD
#> 1      NA     NA  NA          NA            NA        NA          NA
#> 2      NA     NA  NA          NA            NA        NA          NA
#> 3      NA     NA  NA          NA            NA        NA          NA
#> 4      NA     NA  NA          NA            NA        NA          NA
#> 5      NA     NA  NA          NA            NA        NA          NA
#> 6      NA     NA  NA          NA            NA        NA          NA
#>   ClumpingIndex gammaSWR alphaSWR kPAR  g Tmax_LAI Tmax_LAIsq Psi_Extract
#> 1            NA       NA       NA   NA NA       NA         NA          NA
#> 2            NA       NA       NA   NA NA       NA         NA          NA
#> 3            NA       NA       NA   NA NA       NA         NA          NA
#> 4            NA       NA       NA   NA NA       NA         NA          NA
#> 5            NA       NA       NA   NA NA       NA         NA          NA
#> 6            NA       NA       NA   NA NA       NA         NA          NA
#>   Exp_Extract WUE WUE_par WUE_co2 WUE_vpd Gswmin Gswmax Gs_Toptim Gs_Tsens
#> 1          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 2          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 3          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 4          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 5          NA  NA      NA      NA      NA     NA     NA        NA       NA
#> 6          NA  NA      NA      NA      NA     NA     NA        NA       NA
#>   Gs_P50 Gs_slope VCleaf_kmax VCleaf_P12 VCleaf_P50 VCleaf_P88 VCleaf_slope
#> 1     NA       NA          NA         NA         NA         NA           NA
#> 2     NA       NA          NA         NA         NA         NA           NA
#> 3     NA       NA          NA         NA         NA         NA           NA
#> 4     NA       NA          NA         NA         NA         NA           NA
#> 5     NA       NA          NA         NA         NA         NA           NA
#> 6     NA       NA          NA         NA         NA         NA           NA
#>   Kmax_stemxylem VCstem_P12 VCstem_P50 VCstem_P88 VCstem_slope Kmax_rootxylem
#> 1             NA         NA         NA         NA           NA             NA
#> 2             NA         NA         NA         NA           NA             NA
#> 3             NA         NA         NA         NA           NA             NA
#> 4             NA         NA         NA         NA           NA             NA
#> 5             NA         NA         NA         NA           NA             NA
#> 6             NA         NA         NA         NA           NA             NA
#>   VCroot_P12 VCroot_P50 VCroot_P88 VCroot_slope Vmax298 Jmax298 Nleaf Nsapwood
#> 1         NA         NA         NA           NA      NA      NA    NA       NA
#> 2         NA         NA         NA           NA      NA      NA    NA       NA
#> 3         NA         NA         NA           NA      NA      NA    NA       NA
#> 4         NA         NA         NA           NA      NA      NA    NA       NA
#> 5         NA         NA         NA           NA      NA      NA    NA       NA
#> 6         NA         NA         NA           NA      NA      NA    NA       NA
#>   Nfineroot WoodC RERleaf RERsapwood RERfineroot CCleaf CCsapwood CCfineroot
#> 1        NA    NA      NA         NA          NA     NA        NA         NA
#> 2        NA    NA      NA         NA          NA     NA        NA         NA
#> 3        NA    NA      NA         NA          NA     NA        NA         NA
#> 4        NA    NA      NA         NA          NA     NA        NA         NA
#> 5        NA    NA      NA         NA          NA     NA        NA         NA
#> 6        NA    NA      NA         NA          NA     NA        NA         NA
#>   RGRleafmax RGRsapwoodmax RGRcambiummax RGRfinerootmax SRsapwood SRfineroot
#> 1         NA            NA            NA             NA        NA         NA
#> 2         NA            NA            NA             NA        NA         NA
#> 3         NA            NA            NA             NA        NA         NA
#> 4         NA            NA            NA             NA        NA         NA
#> 5         NA            NA            NA             NA        NA         NA
#> 6         NA            NA            NA             NA        NA         NA
#>   RSSG MortalityBaselineRate SurvivalModelStep SurvivalB0 SurvivalB1
#> 1   NA                    NA                NA         NA         NA
#> 2   NA                    NA                NA         NA         NA
#> 3   NA                    NA                NA         NA         NA
#> 4   NA                    NA                NA         NA         NA
#> 5   NA                    NA                NA         NA         NA
#> 6   NA                    NA                NA         NA         NA
#>   SeedProductionHeight SeedMass SeedLongevity DispersalDistance DispersalShape
#> 1                   NA       NA            NA                NA             NA
#> 2                   NA       NA            NA                NA             NA
#> 3                   NA       NA            NA                NA             NA
#> 4                   NA       NA            NA                NA             NA
#> 5                   NA       NA            NA                NA             NA
#> 6                   NA       NA            NA                NA             NA
#>   ProbRecr MinTempRecr MinMoistureRecr MinFPARRecr RecrTreeDBH RecrTreeHeight
#> 1       NA          NA              NA          NA          NA             NA
#> 2       NA          NA              NA          NA          NA             NA
#> 3       NA          NA              NA          NA          NA             NA
#> 4       NA          NA              NA          NA          NA             NA
#> 5       NA          NA              NA          NA          NA             NA
#> 6       NA          NA              NA          NA          NA             NA
#>   RecrShrubHeight RecrTreeDensity RecrShrubCover RecrZ50 RecrZ95 RespFire
#> 1              NA              NA             NA      NA      NA       NA
#> 2              NA              NA             NA      NA      NA       NA
#> 3              NA              NA             NA      NA      NA       NA
#> 4              NA              NA             NA      NA      NA       NA
#> 5              NA              NA             NA      NA      NA       NA
#> 6              NA              NA             NA      NA      NA       NA
#>   RespDist RespClip IngrowthTreeDensity IngrowthTreeDBH
#> 1       NA       NA                  NA              NA
#> 2       NA       NA                  NA              NA
#> 3       NA       NA                  NA              NA
#> 4       NA       NA                  NA              NA
#> 5       NA       NA                  NA              NA
#> 6       NA       NA                  NA              NA

# }