Populate tree species allometries

Populates allometric coefficients for tree species and genus of an input parameter table on the basis of their accepted name.

populate_tree_allometries(
  SpParams,
  allom_table,
  allom_type = "foliarbiomass",
  erase_previous = FALSE
)

populate_shrub_allometries_from_medfuels(
  SpParams,
  coef_mapping,
  sp_params_allom,
  group_params_allom,
  species_groups
)

Arguments

SpParams

A data frame of medfate species parameters to be populated

allom_table

A data frame of allometric parameters in columns and taxonomic entities (species or genus) as row names.

allom_type

A string with the type of allometry to be filled, either "foliarbiomass", "barkthickness", "crownwidth" or "crownratio".

erase_previous

A boolean flag to indicate that values should be set to NA before populating with data

coef_mapping

A named string vector specifying which data column should used to populate each medfate param. Elements are data base columns and names are medfate params.

sp_params_allom

A data table of species allometric coefficients (typically from package medfuels)

group_params_allom

A data table of group allometric coefficients (typically from package medfuels)

species_groups

A data table specifying raunkiaer forms for many species (typically from package medfuels)

Value

A modified data frame of medfate species parameters

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF