Populate species parameters from inventory data

Functions to populate species parameter values from forest inventory source data

populate_GrowthForm(
  SpParams,
  tree_names,
  shrub_names = character(0),
  erase_previous = FALSE,
  fill_fromGenus = FALSE
)

populate_height_params(
  SpParams,
  species_names,
  height_values,
  quantile_Hmed = 0.5,
  quantile_Hmax = 0.99,
  erase_previous = FALSE
)

populate_tree_diameterHeight_params(
  SpParams,
  species_names,
  height_values,
  diameter_values,
  quantile_fHDmin = 0.05,
  quantile_fHDmax = 0.95,
  erase_previous = FALSE
)

Arguments

SpParams

A data frame of medfate species parameters to be populated

tree_names

String vector of translated tree species names

shrub_names

String vector of translated shrub species names

erase_previous

A boolean flag to indicate that values should be set to NA before populating with data

fill_fromGenus

A boolean flag to indicate that genus adscription of species should be used to fill missing values

species_names

A string vector of translated species names (see translateSpeciesCodes)

height_values

A numeric vector of plant heights (in cm)

quantile_Hmed

Quantile for Hmed

quantile_Hmax

Quantile for Hmax

diameter_values

A numeric vector of tree diameter at breast height (in cm)

quantile_fHDmin

Quantile for fHDmin

quantile_fHDmax

Quantile for fHDmax

Value

A modified data frame of medfate species parameters

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF